+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # `gitstat` Changelog
 
-## [0.1] UNRELEASED
+## [0.1.0] UNRELEASED
 
 ### Added
 
@@ -8,7 +8,9 @@
 * Add a fair amount of type hinting
 * Add 'is-tracked' command
 * setuptools packaging
-* Add --version
+* Add '--version'
+* Add '--include-ignored' to several commands
+* Add '--include-existing' to 'showclone'
 
 ### Changed
 
@@ -17,6 +19,7 @@
 * Progress bar now uses tqdm
 * For a few commands, replaced '--quiet' with '--progress'
 * Improve help text
+* '--all' now only available for the 'check' command
 
 
 ## Initial version